+++ RELEASES.md
@@ -6,8 +6,8 @@
   - `GetModifiedAccountsByNumber`
   - `GetModifiedAccountsByHash`
 - Removed deprecated flag `tx-lookup-limit`. Use `transaction-history` instead.
+- Enabled RPC batch limits by default, and configurable with `batch-request-limit` and `batch-max-response-size`.
 - Add pending releases here
-
 
 ## [v0.7.8](https://github.com/ava-labs/subnet-evm/releases/tag/v0.7.8)